+++ pkg/ccl/multiregionccl/BUILD.bazel
@@ -61,6 +61,7 @@ go_test(
         "//pkg/kv/kvpb",
         "//pkg/kv/kvserver",
         "//pkg/kv/kvserver/allocator/allocatorimpl",
+        "//pkg/kv/kvserver/closedts",
         "//pkg/roachpb",
         "//pkg/rpc",
         "//pkg/security/securityassets",
@@ -100,6 +101,7 @@ go_test(
         "//pkg/testutils/testcluster",
         "//pkg/util",
         "//pkg/util/ctxgroup",
+        "//pkg/util/hlc",
         "//pkg/util/leaktest",
         "//pkg/util/log",
         "//pkg/util/quotapool",
