+++ pkg/kv/kvserver/client_split_test.go
@@ -3593,141 +3593,6 @@ func TestSplitBlocksReadsToRHS(t *testing.T) {
 	require.Nil(t, g.Wait())
 }
 
-// TestStoreRangeSplitAndMergeWithGlobalReads tests that a range configured to
-// serve global reads can be split and merged. In essence, this tests whether
-// the split and merge transactions can handle having their timestamp bumped by
-// the closed timestamp on the ranges they're operating on.
-func TestStoreRangeSplitAndMergeWithGlobalReads(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	defer log.Scope(t).Close(t)
-
-	skip.WithIssue(t, 119230)
-
-	// Detect splits and merges over the global read ranges. Assert that the split
-	// and merge transactions commit with pushed write timestamps, and that the
-	// commit-wait sleep for these transactions is performed before running their
-	// commit triggers instead of run on the kv client. For details on why this is
-	// necessary, see maybeCommitWaitBeforeCommitTrigger.
-	var clockPtr atomic.Pointer[hlc.Clock]
-	var splits, merges int64
-	respFilter := func(ctx context.Context, ba *kvpb.BatchRequest, br *kvpb.BatchResponse) *kvpb.Error {
-		clock := clockPtr.Load()
-		if clock == nil {
-			return nil
-		}
-		if req, ok := ba.GetArg(kvpb.EndTxn); ok {
-			endTxn := req.(*kvpb.EndTxnRequest)
-			if br.Txn.Status == roachpb.COMMITTED && br.Txn.MinTimestamp.Less(br.Txn.WriteTimestamp) {
-				if ct := endTxn.InternalCommitTrigger; ct != nil {
-					// The server-side commit-wait sleep should ensure that the commit
-					// triggers are only run after the commit timestamp is below present
-					// time.
-					now := clock.Now()
-					require.True(t, br.Txn.WriteTimestamp.Less(now))
-
-					switch {
-					case ct.SplitTrigger != nil:
-						atomic.AddInt64(&splits, 1)
-					case ct.MergeTrigger != nil:
-						atomic.AddInt64(&merges, 1)
-					}
-				}
-			}
-		}
-		return nil
-	}
-	// Set global reads.
-	zoneConfig := zonepb.DefaultZoneConfig()
-	zoneConfig.GlobalReads = proto.Bool(true)
-
-	ctx := context.Background()
-	s, sqlDB, _ := serverutils.StartServer(t, base.TestServerArgs{
-		Knobs: base.TestingKnobs{
-			Server: &server.TestingKnobs{
-				DefaultZoneConfigOverride: &zoneConfig,
-			},
-
-			Store: &kvserver.StoreTestingKnobs{
-				DisableMergeQueue:     true,
-				TestingResponseFilter: respFilter,
-			},
-		},
-	})
-
-	defer s.Stopper().Stop(ctx)
-	// Set the closed_timestamp interval to be short to shorten the test duration
-	// because we need to wait for a checkpoint on the system config.
-	tdb := sqlutils.MakeSQLRunner(sqlDB)
-	tdb.Exec(t, `SET CLUSTER SETTING kv.closed_timestamp.target_duration = '20ms'`)
-	tdb.Exec(t, `SET CLUSTER SETTING kv.closed_timestamp.side_transport_interval = '20ms'`)
-	tdb.Exec(t, `SET CLUSTER SETTING kv.rangefeed.closed_timestamp_refresh_interval = '20ms'`)
-	store, err := s.GetStores().(*kvserver.Stores).GetStore(s.GetFirstStoreID())
-	require.NoError(t, err)
-	config.TestingSetupZoneConfigHook(s.Stopper())
-
-	// Split off the range for the test.
-	descID := bootstrap.TestingUserDescID(0)
-	descKey := keys.SystemSQLCodec.TablePrefix(descID)
-	splitArgs := adminSplitArgs(descKey)
-	_, pErr := kv.SendWrapped(ctx, store.TestSender(), splitArgs)
-	require.Nil(t, pErr)
-
-	// Set the clock to the store's clock, which also serves to engage the
-	// response filter.
-	clockPtr.Store(s.Clock())
-
-	// Perform a write to the system config span being watched by
-	// the SystemConfigProvider.
-	tdb.Exec(t, "CREATE TABLE foo ()")
-	testutils.SucceedsSoon(t, func() error {
-		repl := store.LookupReplica(roachpb.RKey(descKey))
-		if repl.ClosedTimestampPolicy() != roachpb.LEAD_FOR_GLOBAL_READS {
-			return errors.Errorf("expected LEAD_FOR_GLOBAL_READS policy")
-		}
-		return nil
-	})
-
-	// The commit wait count is 1 due to the split above since global reads are
-	// set for the default config.
-	var splitCount = int64(1)
-	testutils.SucceedsSoon(t, func() error {
-		if splitCount != store.Metrics().CommitWaitsBeforeCommitTrigger.Count() {
-			return errors.Errorf("commit wait count is %d", store.Metrics().CommitWaitsBeforeCommitTrigger.Count())
-		}
-		if splitCount != atomic.LoadInt64(&splits) {
-			return errors.Errorf("num splits is %d", atomic.LoadInt64(&splits))
-		}
-		return nil
-	})
-
-	// Write to the range, which has the effect of bumping the closed timestamp.
-	pArgs := putArgs(descKey, []byte("foo"))
-	_, pErr = kv.SendWrapped(ctx, store.TestSender(), pArgs)
-	require.Nil(t, pErr)
-
-	// Split the range. Should succeed.
-	splitKey := append(descKey, []byte("split")...)
-	splitArgs = adminSplitArgs(splitKey)
-	_, pErr = kv.SendWrapped(ctx, store.TestSender(), splitArgs)
-	require.Nil(t, pErr)
-	splitCount++
-	require.Equal(t, splitCount, store.Metrics().CommitWaitsBeforeCommitTrigger.Count())
-	require.Equal(t, splitCount, atomic.LoadInt64(&splits))
-
-	repl := store.LookupReplica(roachpb.RKey(splitKey))
-	require.Equal(t, splitKey, repl.Desc().StartKey.AsRawKey())
-
-	// Merge the range. Should succeed.
-	mergeArgs := adminMergeArgs(descKey)
-	_, pErr = kv.SendWrapped(ctx, store.TestSender(), mergeArgs)
-	require.Nil(t, pErr)
-	require.Equal(t, splitCount+1, store.Metrics().CommitWaitsBeforeCommitTrigger.Count())
-	require.Equal(t, int64(1), atomic.LoadInt64(&merges))
-
-	repl = store.LookupReplica(roachpb.RKey(splitKey))
-	require.Equal(t, descKey, repl.Desc().StartKey.AsRawKey())
-}
-
 // TestLBSplitUnsafeKeys tests that load based splits do not split between table
 // rows, even when the suggested load based split key is itself between a table row.
 func TestLBSplitUnsafeKeys(t *testing.T) {
